refactor: isolate tmux store persistence
